requirements.txt
requests
playwright
sentence-transformers
faiss-cpu
numpy
sseclient-py 